requirements.txt
streamlit>=1.32.0
onnxruntime>=1.17.0
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
plotly>=5.18.0
scikit-learn>=1.3.0
fpdf2>=2.7.6
requests>=2.31.0
streamlit-extras>=0.3.6
